Update 3Dstuff.py
This program measures the angle of the beagle bone in relation to its x, y, and z axes. the angle used to determine the robots angle around its axes is beta. The code will print whether the robot is facing up or down a slope and record the 2D component of its distance traveled.